Adds receipt generator and exception handlder

+++ src/main/java/nl/bd/garage/services/EmployeeService.java
@@ -1,6 +1,8 @@
 package nl.bd.garage.services;
 
 import nl.bd.garage.models.entities.Employee;
+import nl.bd.garage.models.enums.Role;
+import nl.bd.garage.models.exceptions.ModifyAdminException;
 import nl.bd.garage.models.exceptions.EmployeeNotFoundException;
 import nl.bd.garage.repositories.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,25 +28,37 @@ public class EmployeeService {
     }
 
     public Employee createEmployee(Employee newEmployee) {
-        newEmployee.setPassword(passwordEncoder.encode(newEmployee.getPassword()));
-        return employeeRepository.save(newEmployee);
+        if (newEmployee.getRole() == Role.ADMIN) {
+            throw new ModifyAdminException("create");
+        } else {
+            newEmployee.setPassword(passwordEncoder.encode(newEmployee.getPassword()));
+            return employeeRepository.save(newEmployee);
+        }
     }
 
     public Employee updateEmployee(Employee newEmployee, Long employeeId) {
-        return employeeRepository.findById(employeeId)
-                .map(employee -> {
-                    employee.setUsername(newEmployee.getUsername());
-                    employee.setPassword(passwordEncoder.encode(newEmployee.getPassword()));
-                    employee.setName(newEmployee.getName());
-                    employee.setRole(newEmployee.getRole());
-                    return employeeRepository.save(employee);
-                })
-                .orElseThrow(() -> new EmployeeNotFoundException(employeeId));
+        if (newEmployee.getRole() == Role.ADMIN) {
+            throw new ModifyAdminException("create");
+        } else {
+            return employeeRepository.findById(employeeId)
+                    .map(employee -> {
+                        employee.setUsername(newEmployee.getUsername());
+                        employee.setPassword(passwordEncoder.encode(newEmployee.getPassword()));
+                        employee.setName(newEmployee.getName());
+                        employee.setRole(newEmployee.getRole());
+                        return employeeRepository.save(employee);
+                    })
+                    .orElseThrow(() -> new EmployeeNotFoundException(employeeId));
+        }
     }
 
     public void deleteEmployee(Long employeeId) {
         if (employeeRepository.findById(employeeId).isPresent()) {
-            employeeRepository.deleteById(employeeId);
+            if (employeeRepository.findById(employeeId).get().getRole() == Role.ADMIN) {
+                throw new ModifyAdminException("delete");
+            } else {
+                employeeRepository.deleteById(employeeId);
+            }
         } else {
             throw new EmployeeNotFoundException(employeeId);
         }
